TestTask/requirements.txt
opencv-python
Pillow
albumentations
numpy
albumentations
matplotlib